fix: apply -allow-git flag even when using presets
