+++ dryrun_linux.go
@@ -4,7 +4,6 @@ package main
 
 import (
 	"fmt"
-	"os"
 	"path/filepath"
 	"strings"
 )
@@ -21,7 +20,7 @@ func showDryRun(config *SandboxConfig) error {
 	fmt.Println("Rules:")
 
 	if config.AllowAll {
-		fmt.Println("- Allow all operations (--allow-all flag)")
+		fmt.Println("- Allow all operations (-allow-all flag)")
 	} else {
 		fmt.Println("- Allow read access to all files")
 		fmt.Println("- Deny write access except to:")
@@ -35,7 +34,7 @@ func showDryRun(config *SandboxConfig) error {
 			}
 			source := "user specified"
 			if config.AllowGit && strings.Contains(path, ".git") {
-				source = "--allow-git"
+				source = "-allow-git"
 			}
 			fmt.Printf("  * %s (%s)\n", absPath, source)
 		}
@@ -49,13 +48,4 @@ func showDryRun(config *SandboxConfig) error {
 	fmt.Println()
 
 	return nil
-}
-
-// printDryRunAndExit displays the dry-run information and exits
-func printDryRunAndExit(config *SandboxConfig) {
-	if err := showDryRun(config); err != nil {
-		fmt.Fprintf(os.Stderr, "cage: error showing dry-run: %v\n", err)
-		os.Exit(1)
-	}
-	os.Exit(0)
 }
