+++ main.go
@@ -4,7 +4,6 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"path/filepath"
 	"runtime/debug"
 	"strings"
 )
@@ -189,10 +188,6 @@ func main() {
 	allowKeychain := flags.allowKeychain
 	allowGit := flags.allowGit
 
-	// Track unique paths to avoid duplicates
-	pathSet := make(map[string]struct{})
-	var uniquePaths []string
-
 	// Process each preset and merge their settings
 	for _, presetName := range flags.presets {
 		preset, ok := config.GetPreset(presetName)
@@ -208,17 +203,9 @@ func main() {
 			os.Exit(1)
 		}
 
-		// Add preset paths, checking for duplicates
+		// Add preset paths
 		for _, path := range processedPreset.Allow {
-			absPath, err := filepath.Abs(path.Path)
-			if err != nil {
-				// If we can't get absolute path, use original path
-				absPath = path.Path
-			}
-			if _, exists := pathSet[absPath]; !exists {
-				pathSet[absPath] = struct{}{}
-				uniquePaths = append(uniquePaths, absPath)
-			}
+			allowedPaths = append(allowedPaths, path.Path)
 		}
 
 		// Preset's allowKeychain is ORed with command-line flag
@@ -226,33 +213,6 @@ func main() {
 
 		// Preset's allowGit is ORed with command-line flag
 		allowGit = allowGit || processedPreset.AllowGit
-	}
-
-	// Add command-line paths, checking for duplicates
-	for _, path := range allowedPaths {
-		absPath, err := filepath.Abs(path)
-		if err != nil {
-			// If we can't get absolute path, use original path
-			absPath = path
-		}
-		if _, exists := pathSet[absPath]; !exists {
-			pathSet[absPath] = struct{}{}
-			uniquePaths = append(uniquePaths, path)
-		}
-	}
-
-	// Replace allowedPaths with unique paths
-	allowedPaths = uniquePaths
-
-	// Add git common directory if allowGit is enabled and not already handled by preset
-	if allowGit && len(flags.presets) == 0 {
-		gitCommonDir, err := getGitCommonDir()
-		if err != nil {
-			// Log the error but don't fail - the directory might not be a git repo
-			fmt.Fprintf(os.Stderr, "warning: %v\n", err)
-		} else {
-			allowedPaths = append(allowedPaths, gitCommonDir)
-		}
 	}
 
 	// Create sandbox configuration
